refactor(update-tokens): extract squid-api module and parallelize color extraction
Splits the Squid API client out of colors.js so it can be reused without
side effects. Previously, importing `getSquidAssets` from colors.js
triggered its top-level IIFE, causing the full color-extraction pass
(and Squid API fetch) to run twice on every `yarn update-tokens`.

- New `squid-api.js` library: `getSquidAssets` + helpers, zero side effects.
- `colors.js` is now a pure entry script — no exports, with a runtime
  guard that throws if anything tries to import it. fetch-new-tokens.js
  gets the same guard.
- Token color loop now runs in batches of `TOKEN_CONCURRENCY = 16`
  instead of sequentially with a 100ms sleep per token.

The duplicate-run cost on the previous structure was a full second pass
of color extraction; the new parallel loop also cuts the remaining work
substantially.
