+++ scripts/update-tokens/colors.js
@@ -1,111 +1,32 @@
+// ENTRY SCRIPT — do not import this file from anywhere.
+// This runs `main()` at module load; importing it would trigger an unintended
+// full color-extraction pass (and re-fetch all Squid data).
+// Shared helpers live in `./squid-api.js` and `./colors-utils.js`.
+
 import "dotenv/config"
 import chalk from "chalk"
 import fs from "node:fs"
+import { pathToFileURL } from "node:url"
 import {
   getTokenAssetsKey,
   getAverageColor,
-  getContrastColor,
-  isEvmosChain,
-  isSolanaSanctumAutomatedToken,
-  nativeEvmTokenAddress
+  getContrastColor
 } from "./colors-utils.js"
+import { getSquidAssets } from "./squid-api.js"
+
+if (import.meta.url !== pathToFileURL(process.argv[1]).href) {
+  throw new Error(
+    "colors.js is an entry script and must not be imported. " +
+      "Import shared helpers from ./squid-api.js or ./colors-utils.js instead."
+  )
+}
+
 const colorsFilePath = "scripts/update-tokens/colors.json"
 const failedUrlsFilePath = "scripts/update-tokens/url_fetch_errors.json"
 const defaultChainBgColor = ""
 const defaultTokenBgColor = ""
 const defaultTokenTextColor = ""
-
-const sleep = ms => new Promise(resolve => setTimeout(resolve, ms))
-
-export const getSquidAssets = async (retries = 3) => {
-  for (let attempt = 1; attempt <= retries; attempt++) {
-    try {
-      console.log(`Attempting to fetch Squid data (attempt ${attempt}/${retries})...`)
-      
-      const url = new URL("/v2/sdk-info", process.env.SQUID_API_URL)
-      
-      // Add timeout to the fetch request
-      const controller = new AbortController()
-      const timeoutId = setTimeout(() => controller.abort(), 30000) // 30 second timeout
-      
-      const response = await fetch(url, {
-        headers: {
-          "Content-Type": "application/json",
-          "x-integrator-id": process.env.SQUID_INTEGRATOR_ID
-        },
-        signal: controller.signal
-      })
-
-      clearTimeout(timeoutId)
-
-      if (!response.ok) {
-        throw new Error(`HTTP error! status: ${response.status}`)
-      }
-
-      const data = await response.json()
-      validateSdkInfo(data)
-
-      // remove sanctum automated tokens
-      data.tokens = data.tokens.filter(t => !isSolanaSanctumAutomatedToken(t))
-
-      const evmosChains = data.chains.filter(isEvmosChain)
-      const evmosChainIds = evmosChains.map(c => c.chainId)
-
-      const evmosNativeTokenAddressesMap = evmosChains.reduce((acc, chain) => {
-        const normalizedSymbol = chain.nativeCurrency.symbol.toLowerCase()
-
-        return {
-          ...acc,
-          [normalizedSymbol]: nativeEvmTokenAddress
-        }
-      }, {})
-
-      /**
-       * Converts an evmos address (erc20/0x123...abc)
-       * to an evm standard address (0x123...abc)
-       *
-       * Also gas tokens on evmos chains have non-standard EVM addresses
-       * so we need to map them to the native EVM token address
-       */
-      const evmosAddressToEvmAddress = address => {
-        if (evmosNativeTokenAddressesMap[address]) {
-          return evmosNativeTokenAddressesMap[address]
-        }
-
-        return address.replace(/^erc20\//, "")
-      }
-
-      data.tokens = data.tokens.map(token => {
-        const isEvmosToken = evmosChainIds.includes(token.chainId)
-
-        return {
-          ...token,
-          address: isEvmosToken
-            ? // convert evmos address (erc20/0x123...abc) to evm address (0x123...abc)
-              evmosAddressToEvmAddress(token.address)
-            : token.address
-        }
-      })
-
-      console.log(chalk.green(`Successfully fetched ${data.chains.length} chains and ${data.tokens.length} tokens`))
-      return data
-      
-    } catch (error) {
-      console.error(`Attempt ${attempt} failed:`, error.message)
-      
-      if (attempt === retries) {
-        console.error(chalk.red("All attempts failed. Returning empty data structure."))
-        // Return proper structure instead of empty array
-        return { chains: [], tokens: [] }
-      }
-      
-      // Wait before retrying (exponential backoff)
-      const waitTime = Math.pow(2, attempt) * 1000
-      console.log(chalk.yellow(`Waiting ${waitTime}ms before retry...`))
-      await sleep(waitTime)
-    }
-  }
-}
+const TOKEN_CONCURRENCY = 16
 
 const defaultColors = { tokens: {}, chains: {} }
 const getSavedColors = () => {
@@ -146,18 +67,36 @@ function saveFailedUrls(failedUrls) {
   })
 }
 
-;(async function main() {
+function getRgbKeys(color) {
+  const [r = 0, g = 0, b = 0] = color.match(/\d+/g)
+
+  return {
+    r: Number(r),
+    g: Number(g),
+    b: Number(b)
+  }
+}
+
+// Use png instead of webp, as webp is not supported by node canvas
+const getTokenImage = token =>
+  `images/migration/png/${getTokenAssetsKey(token)}.png`
+
+const getChainImage = chain => {
+  return chain.chainIconURI.replaceAll("webp", "png")
+}
+
+async function main() {
   console.log("Extracting assets colors")
 
   const squidData = await getSquidAssets()
-  
+
   // Add safety checks for destructuring
   const { tokens = [], chains = [] } = squidData || {}
-  
+
   if (chains.length === 0) {
     console.log(chalk.yellow("No chains data available. Skipping chain processing."))
   }
-  
+
   if (tokens.length === 0) {
     console.log(chalk.yellow("No tokens data available. Skipping token processing."))
   }
@@ -177,11 +116,6 @@ function saveFailedUrls(failedUrls) {
 
   for (const chain of chains) {
     if (!!colors.chains[chain.chainId]?.bgColor) {
-      // console.log(
-      //   chalk.grey(
-      //     `Chain ${chain.networkName} (${chain.chainId}) already exists, skipping`
-      //   )
-      // )
       continue
     }
 
@@ -223,29 +157,18 @@ function saveFailedUrls(failedUrls) {
 
   await Promise.all(chainColorPromises)
 
-  for await (const token of tokens) {
-    // add chain object if it doesn't exist
+  // Ensure every chain referenced by a token has at least a default entry
+  for (const token of tokens) {
     if (!colors.chains[token.chainId]) {
-      colors.chains[token.chainId] = {
-        bgColor: defaultChainBgColor
-      }
+      colors.chains[token.chainId] = { bgColor: defaultChainBgColor }
     }
+  }
 
-    const tokenAlreadyExists =
-      !!colors.tokens[getTokenAssetsKey(token)]?.bgColor
+  const tokensToProcess = tokens.filter(
+    token => !colors.tokens[getTokenAssetsKey(token)]?.bgColor
+  )
 
-    if (tokenAlreadyExists) {
-      // console.log(
-      //   chalk.grey(
-      //     `Token ${token.symbol} on ${
-      //       chainIdToNameMapping[token.chainId]
-      //     } already exists, skipping`
-      //   )
-      // )
-      continue
-    }
-
-    await sleep(100)
+  const processToken = async token => {
     try {
       const tokenBgColor = await getAverageColor(getTokenImage(token))
       const { r, g, b } = getRgbKeys(tokenBgColor)
@@ -256,31 +179,21 @@ function saveFailedUrls(failedUrls) {
         textColor: tokenTextColor
       }
       console.log(
-        chalk.rgb(
-          r,
-          g,
-          b
-        )(
-          `Token ${token.symbol} on ${
-            chainIdToNameMapping[token.chainId]
-          } saved`
+        chalk.rgb(r, g, b)(
+          `Token ${token.symbol} on ${chainIdToNameMapping[token.chainId]} saved`
         )
       )
     } catch (error) {
       console.error(
         chalk.bgRed.white.underline.bold(
-          `Error fetching image for token ${token.symbol} on ${
-            chainIdToNameMapping[token.chainId]
-          }:`
+          `Error fetching image for token ${token.symbol} on ${chainIdToNameMapping[token.chainId]}:`
         ),
         error.message
       )
       console.log("at", chalk.blueBright(token.logoURI))
       console.log(
         chalk.grey(
-          `Token ${token.symbol} on ${
-            chainIdToNameMapping[token.chainId]
-          } saved using fallback colors`
+          `Token ${token.symbol} on ${chainIdToNameMapping[token.chainId]} saved using fallback colors`
         ),
         "\n"
       )
@@ -299,29 +212,12 @@ function saveFailedUrls(failedUrls) {
     }
   }
 
+  for (let i = 0; i < tokensToProcess.length; i += TOKEN_CONCURRENCY) {
+    const batch = tokensToProcess.slice(i, i + TOKEN_CONCURRENCY)
+    await Promise.all(batch.map(processToken))
+  }
+
   await Promise.all([saveColors(colors), saveFailedUrls(failedUrls)])
-})()
-
-export function getRgbKeys(color) {
-  const [r = 0, g = 0, b = 0] = color.match(/\d+/g)
-
-  return {
-    r: Number(r),
-    g: Number(g),
-    b: Number(b)
-  }
 }
 
-// Use png instead of webp, as webp is not supported by node canvas
-export const getTokenImage = token =>
-  `images/migration/png/${getTokenAssetsKey(token)}.png`
-
-export const getChainImage = chain => {
-  return chain.chainIconURI.replaceAll("webp", "png")
-}
-
-function validateSdkInfo(data) {
-  if (!Array.isArray(data.chains) || !Array.isArray(data.tokens)) {
-    throw new Error("Invalid Squid data: missing chains or tokens")
-  }
-}
+main()
